# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Food Court QR Ordering Platform API
  version: 1.0.0-pilot-STALE
  description: |
    # ⚠ THIS DOCUMENT IS STALE. DO NOT GENERATE A CLIENT FROM IT.

    Superseded by `docs/prd.md` v7.0 and by the running code. It is kept only so
    the divergence is visible; every disagreement below resolves in favour of
    the implementation.

    The warning lives here, in the file itself, rather than only in the PRD.
    Somebody who opens this directly — to generate a client, to check a field
    name, to write an integration — would otherwise get no warning at all, and
    a stale contract read in good faith produces confidently wrong code.

    ## What is known to be wrong

    | Here | Actually |
    | --- | --- |
    | `/qr/{token}` returns a `table` object | It resolves to a FOOD COURT. The table concept was removed — PRD §2.1 |
    | Order status `COMPLETED` | `COLLECTED`. Cooked and handed over are different events — PRD §7.1 |
    | Payment status `SUCCESS` | `AUTHORIZED` and `CAPTURED` are separate states — PRD §8.1 |
    | Menu item `is_available` boolean | `status` / `availability` / `inventory_mode`, three concepts — PRD §6 |
    | `POST /kds/orders/{id}/complete` | `/collect` |
    | no session or OTP endpoints | `/auth/otp/request`, `/auth/otp/verify` exist — PRD §11.2 |
    | no payment endpoints | `/orders/{id}/payment-intent`, `/webhooks/payments`, `/orders/{id}/payment` exist |
...

...
    script produces types that contradict the server.

...

...

...

...

...
servers:
...
security:
  - sessionToken: []
paths:
...
  /orders:
    post:
      tags: [Checkout]
      operationId: createOrder
      summary: Create an order and its payment intent
      description: |
        PRD PAY-01/02, API-01. Requires `Idempotency-Key`. Creates the order in
        `CREATED`, snapshots the fee rule, settlement mode and tax model, then
        creates the provider payment intent and moves the order to
        `PAYMENT_PENDING`. Replaying the same key returns the original order and
        payment intent with `200`.

        This endpoint never confirms payment. Confirmation arrives only through
        the verified provider webhook.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [cartId, quoteId, customerPhone]
              properties:
                cartId: { type: string, format: uuid }
                quoteId:
                  {
                    type: string,
                    format: uuid,
                    description: From /checkout/validate. Binds the price.,
                  }
                customerPhone:
                  type: string
                  pattern: '^\+[1-9][0-9]{7,14}$'
                  description: |
                    Captured, not verified. PRD §8.2 — payment success is the
                    implicit verification event. No OTP in the pilot.
                whatsappOptIn: { type: boolean, default: false }
                applyCreditId: { type: string, format: uuid, nullable: true }
      responses:
        '201':
          description: Order created, payment intent issued
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OrderCreated' }
        '200':
          description: Idempotent replay — the original order is returned unchanged.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OrderCreated' }
        '400':
          description: Missing `Idempotency-Key`.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Problem' }
        '409':
          description: Quote expired or checkout no longer valid. Re-run /checkout/validate.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Problem' }

  # ==================================================================
  # PAYMENT
  # ==================================================================
...
  /kds/heartbeat:
    post:
      tags: [KDS]
      operationId: kdsHeartbeat
      security: [{ deviceToken: [] }]
      summary: Device liveness beacon
      description: |
        PRD KDS-HB-01. Sent every 15 seconds. Silence for 60 seconds marks the
        vendor unavailable for new orders and auto-recovers on resume.
        **This endpoint is what earns the right to auto-fulfil** — without it the
        accept click cannot be removed.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [deviceId, appVersion, unacknowledgedCount]
              properties:
                deviceId: { type: string, format: uuid }
                appVersion: { type: string }
                unacknowledgedCount: { type: integer, minimum: 0 }
                printerOnline: { type: boolean, nullable: true }
      responses:
        '200':
          description: Acknowledged, with any server-side directives
          content:
            application/json:
              schema:
                type: object
                properties:
                  serverTime: { type: string, format: date-time }
                  vendorOrderable: { type: boolean }
                  dispatchBlocked:
                    { type: boolean, description: True while an unacknowledged order is past 90s. }
                  pendingOrderIds:
                    type: array
                    items: { type: string, format: uuid }
  /kds/orders:
    get:
      tags: [KDS]
      operationId: kdsListOrders
      security: [{ deviceToken: [] }]
      parameters:
        - name: status
          in: query
          schema:
            type: array
            items: { $ref: '#/components/schemas/OrderStatus' }
      responses:
        '200':
          description: Live queue for this vendor
          content:
            application/json:
              schema:
                type: object
                properties:
                  orders:
                    type: array
                    items: { $ref: '#/components/schemas/KdsOrder' }
  /orders/{orderId}/acknowledge:
    post:
      tags: [KDS]
      operationId: acknowledgeOrder
      security: [{ deviceToken: [] }]
      summary: Machine confirmation that the order was rendered or printed
      description: |
        PRD §9, KDS-PR-02. **This is not an accept action and no human triggers
        it.** The KDS calls it automatically once the order card has painted;
        the print worker calls it on a successful print. It cancels the
        escalation ladder. A print failure must NOT call this endpoint.
      parameters:
        - { $ref: '#/components/parameters/OrderId' }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [targetType, deviceId]
              properties:
                targetType: { $ref: '#/components/schemas/DispatchTarget' }
                deviceId: { type: string, format: uuid }
      responses:
        '200':
          description: Acknowledged; ladder cancelled
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Order' }
        '409': { $ref: '#/components/responses/InvalidTransition' }
  /orders/{orderId}/preparing:
    post:
      tags: [KDS]
      operationId: markPreparing
      security: [{ deviceToken: [] }]
      parameters: [{ $ref: '#/components/parameters/OrderId' }]
      responses:
        '200':
          description: Marked preparing
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Order' }
        '409': { $ref: '#/components/responses/InvalidTransition' }
...
  /orders/{orderId}/complete:
    post:
      tags: [KDS]
      operationId: markCompleted
      security: [{ deviceToken: [] }]
      parameters: [{ $ref: '#/components/parameters/OrderId' }]
      responses:
        '200':
          description: Marked completed
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Order' }
        '409': { $ref: '#/components/responses/InvalidTransition' }
  /orders/{orderId}/reject:
    post:
      tags: [KDS]
      operationId: rejectOrder
      security: [{ deviceToken: [] }]
      summary: Exception path — reject and auto-refund
      description: |
        PRD KDS-REJ-01..05. Reason is mandatory. Atomically blocks all further
        fulfilment transitions and creates the refund command with no human step
        (REF-01, within 60 seconds). Available from `DISPATCHED` through
        `PREPARING`; disabled once `READY`.
      parameters:
        - { $ref: '#/components/parameters/OrderId' }
        - { $ref: '#/components/parameters/IdempotencyKey' }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reason]
              properties:
                reason: { $ref: '#/components/schemas/RejectionReason' }
                note: { type: string, maxLength: 300, nullable: true }
                rejectedItemIds:
                  type: array
                  description: V1 only. Item-level rejection with partial refund (KDS-REJ-06).
                  items: { type: string, format: uuid }
      responses:
        '200':
          description: Rejected; refund initiated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Order' }
        '409': { $ref: '#/components/responses/InvalidTransition' }
...
components:
  securitySchemes:
    sessionToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Short-lived customer session token issued by POST /sessions.
    deviceToken:
      type: http
      scheme: bearer
      description: KDS or printer device token. Scoped to one vendor.
    staffToken:
      type: http
      scheme: bearer
      description: Manager, ops, finance and admin. Role and tenant scope are claims.
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: |
        Client-generated, unique per attempt. Backed by a unique database
        constraint. Replay returns the original result. PRD API-01.
      schema: { type: string, minLength: 16, maxLength: 128 }
...
    OrderId:
      name: orderId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
...
    InvalidTransition:
      description: |
        The requested transition is not legal from the order's current state.
        Never a silent no-op and never a forced write (PRD ORD-SM-01).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Problem' }
  schemas:
    Problem:
      type: object
      required: [code, message, correlationId]
      properties:
        code:
          type: string
          description: |
            Stable machine-readable error code. Clients branch on this, never on
            the message. See TDD §18 for the full catalogue.
          examples:
            [VENDOR_DEVICE_OFFLINE, INVALID_TRANSITION, QUOTE_EXPIRED, IDEMPOTENCY_KEY_REQUIRED]
        message: { type: string, description: Human-readable, not for display to customers. }
        detail: { type: string, nullable: true }
        correlationId: { type: string }
        fields:
          type: object
          nullable: true
          additionalProperties: { type: string }
    Money:
      type: integer
      format: int64
      minimum: 0
      description: Amount in paise. 25000 = ₹250.00. Never a decimal.
    OrderStatus:
      type: string
      enum:
        [
          CREATED,
          PAYMENT_PENDING,
          PAYMENT_CONFIRMED,
          DISPATCHED,
          ACKNOWLEDGED,
          PREPARING,
          READY,
          COMPLETED,
          PAYMENT_FAILED,
          DISPATCH_FAILED,
          REJECTED,
          CANCELLED,
          REFUND_PENDING,
          REFUNDED,
          REFUND_FAILED,
          RECONCILIATION_REQUIRED,
        ]
...
    DispatchTarget:
      type: string
      enum: [KDS, PRINTER, POS]
    RejectionReason:
      type: string
      enum:
        [
          ITEM_OUT_OF_STOCK,
          KITCHEN_OVERLOADED,
          EQUIPMENT_FAILURE,
          VENDOR_CLOSING,
          INGREDIENT_UNAVAILABLE,
          PRICE_OR_MENU_ERROR,
          CUSTOMER_REQUEST,
          DUPLICATE_ORDER,
          OTHER,
        ]
...
    OrderCreated:
      type: object
      properties:
        order: { $ref: '#/components/schemas/Order' }
        payment:
          type: object
          properties:
            paymentId: { type: string, format: uuid }
            provider: { type: string }
            providerOrderRef: { type: string }
            amountPaise: { $ref: '#/components/schemas/Money' }
            checkoutPayload:
              type: object
              description: Opaque provider payload the client hands to the checkout SDK.
              additionalProperties: true
    Order:
      type: object
      properties:
        id: { type: string, format: uuid }
        publicOrderNumber: { type: string }
        status: { $ref: '#/components/schemas/OrderStatus' }
        foodCourtId: { type: string, format: uuid }
        vendorId: { type: string, format: uuid }
        vendorName: { type: string }
        tableLabel: { type: string }
        items:
          type: array
          items:
            type: object
            properties:
              name: { type: string }
              quantity: { type: integer }
              lineTotalPaise: { $ref: '#/components/schemas/Money' }
              options:
                type: array
                items: { type: string }
              instructions: { type: string, nullable: true }
              isRejected: { type: boolean }
        totalPayablePaise: { $ref: '#/components/schemas/Money' }
        rejectionReason:
          oneOf:
            - { $ref: '#/components/schemas/RejectionReason' }
            - { type: 'null' }
        refund:
          oneOf:
            - { $ref: '#/components/schemas/Refund' }
            - { type: 'null' }
        creditIssued:
          oneOf:
            - { $ref: '#/components/schemas/PlatformCredit' }
            - { type: 'null' }
        cancellationOffered:
          type: boolean
          description: True once the escalation ladder reaches 180s (PRD ESC-03).
        estimatedReadyAt: { type: string, format: date-time, nullable: true }
        timeline:
          type: array
          items:
            type: object
            properties:
              status: { $ref: '#/components/schemas/OrderStatus' }
              at: { type: string, format: date-time }
              actorType: { type: string }
        createdAt: { type: string, format: date-time }
    KdsOrder:
      type: object
      description: |
        PRD KDS-02. Everything the kitchen needs to cook and hand over without
        looking at another screen. Deliberately excludes payment detail.
      properties:
        id: { type: string, format: uuid }
        publicOrderNumber: { type: string }
        tableLabel: { type: string }
        status: { $ref: '#/components/schemas/OrderStatus' }
        placedAt: { type: string, format: date-time }
        ageSeconds: { type: integer }
        slaSeconds: { type: integer }
        totalPayablePaise: { $ref: '#/components/schemas/Money' }
        items:
          type: array
          items:
            type: object
            properties:
              name: { type: string }
              quantity: { type: integer }
              options:
                type: array
                items: { type: string }
              instructions: { type: string, nullable: true }
...
    Refund:
      type: object
      properties:
        id: { type: string, format: uuid }
        orderId: { type: string, format: uuid }
        kind: { type: string, enum: [FULL, PARTIAL] }
        amountPaise: { $ref: '#/components/schemas/Money' }
        status: { type: string, enum: [REQUESTED, PENDING, SUCCEEDED, FAILED, ABANDONED] }
        customerMessage:
          type: string
          description: |
            Explicit customer-facing status. The words "payment lost" or an
            unexplained failure never appear (PRD CUS-TRK-05).
        expectedByAt: { type: string, format: date-time, nullable: true }
        createdAt: { type: string, format: date-time }
    PlatformCredit:
      type: object
      description: |
        PRD REF-05. Issued instantly at rejection so the customer is not left
        waiting on a bank refund. Mutually exclusive with refund-to-source,
        enforced atomically.
      properties:
        id: { type: string, format: uuid }
        amountPaise: { $ref: '#/components/schemas/Money' }
        status: { type: string, enum: [ISSUED, CONSUMED, EXPIRED, VOIDED] }
        expiresAt: { type: string, format: date-time }
...